fix: load modules error on others when they had no modules.nix besides configuartion.nix
