+++ src/lib.rs
@@ -1,6 +1,3 @@
 pub mod config;
 pub mod ui;
 pub mod utils;
-
-const MODULES_CONFIG: &str =
-    "/home/victor/Documents/threadripperwork/nix2/systems/x86_64-linux/xinux/modules.nix";
