+++ src/ui/mod.rs
@@ -19,7 +19,6 @@ pub mod wellbeing;
 pub mod mouse;
 
 pub mod accessibility;
-pub mod load;
 pub mod privacyandsecurity;
 pub mod rebuild;
 pub mod system;
